Fixed warnings and changed premake files

+++ SK-Game-Engine/src/SK-Game-Engine/ImGui/ImGuiLayer.cpp
@@ -42,7 +42,7 @@ namespace SK_Game_Engine
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Window& window = Application::Get().GetWindow();
-		io.DisplaySize = ImVec2(window.GetWidth(), window.GetHeight());
+		io.DisplaySize = ImVec2(static_cast<float>(window.GetWidth()), static_cast<float>(window.GetHeight()));
 		ImGui::Render();
 
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
